Notification property order fixed in manual interface implementations. Transaction Log class created which collects cumulative changes within a scope of transaction. Nuspec++

+++ Xbim.CobieExpress/OptionalItemSet.cs
@@ -14,7 +14,7 @@ namespace Xbim.CobieExpress
     {
         private bool _initialized;
 
-        internal OptionalItemSet(IPersistEntity entity, int capacity, byte property) : base(entity, capacity, property)
+        internal OptionalItemSet(IPersistEntity entity, int capacity, int property) : base(entity, capacity, property)
         {
         }
